requirements.txt
Flask>=2.0
python-dotenv>=0.21
groq>=0.13.0
